requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mongolock==1.3.4
pymongo==3.9.0
python-engineio==3.10.0
python-socketio==4.4.0
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
gevent
